pyproject.toml
[tool.poetry]
name = "repository_manager"
version = "0.1.0"
description = ""
authors = ["Shahin Azad <hi@shahinism.com>"]
readme = "README.md"
packages = [{include = "repository_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
pygithub = "^1.58.0"
python-slugify = "^8.0.1"
gitpython = "^3.1.31"
logzero = "^1.7.0"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
ipython = "^8.11.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
togh = "repository_manager.togh:main"